Add config.el
I am starting to track config.el with the rest of the project, because
it was a pain in the ass when I was making changes to how the config
file gets tangled and then not noticing that my changes were not
applied, because the file was not tangled anymore. It might be a better
idea to handle the tangling seperately, but I am to lazy to do that.