Update org-roam directory
I updated the main directory for org-roam, so that I don't have org-roam
enabled when filing my notes.org file. Things like archiving are way
slower if org-roam is enabled while in the notes.org file because it
want's to update it's database the whole time. That gets really annoying quickly.